+++ lib/app/constants/feedback_options.dart
@@ -1,0 +1,9 @@
+class FeedbackOptions {
+  static const List<String> options = [
+    "Harmful or abusive",
+    "Irrelevant or off topic",
+    "Inappropriate or offensive",
+    "Ending conversation abruptly",
+    "Other",
+  ];
+}
