+++ lib/ui/widgets/messaging/feedback/feedback_box.dart
@@ -1,18 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_chatbot/app/constants/feedback_options.dart';
 import 'package:flutter_chatbot/ui/widgets/messaging/feedback/feedback_item.dart';
 
 class FeedbackBox extends StatelessWidget {
   final Function(int) setFeedbackView;
 
   FeedbackBox({this.setFeedbackView});
-
-  final List<String> options = [
-    "Harmful or abusive",
-    "Irrelevant or off topic",
-    "Inappropriate or offensive",
-    "Ending conversation abruptly",
-    "Other",
-  ];
 
   Widget build(BuildContext context) {
     return Container(
@@ -24,10 +17,10 @@ class FeedbackBox extends StatelessWidget {
         child: ListView.builder(
             shrinkWrap: true,
             padding: EdgeInsets.symmetric(vertical: 15, horizontal: 20),
-            itemCount: options.length,
+            itemCount: FeedbackOptions.options.length,
             itemBuilder: (context, index) {
               return FeedbackItem(
-                  optionText: options[index],
+                  optionText: FeedbackOptions.options[index],
                   optionNum: index + 1,
                   setFeedbackView: setFeedbackView);
             }));
